% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{add_chance_line}
\alias{add_chance_line}
\title{Show chance line in a ROC plot}
\usage{
add_chance_line()
}
\description{
Plot chance line in a ROC plot.
}
\examples{
plot_roc_curve(iris, response = Species, predictor = Sepal.Width) +
 add_chance_line()
}
