% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{calc_curve_shape}
\alias{calc_curve_shape}
\title{Calculate curve shape over an specific region}
\usage{
calc_curve_shape(
  data = NULL,
  response = NULL,
  predictor = NULL,
  lower_threshold,
  upper_threshold,
  ratio
)
}
\arguments{
\item{data}{data.frame}

\item{response}{NULL}

\item{predictor}{NULL}

\item{lower_threshold, upper_threshold}{NULL}

\item{ratio}{NULL}
}
\value{
A character vector indicating ROC curve shape in the specified region which
can take any of the following values:
\itemize{
\item \code{Concave}. ROC curve is concave over the entire region.
\item \verb{Partially proper}. ROC curve loses concavity at some point of the
region.
\item \verb{Hook under chance}. ROC curve loses concavity at some point of the
region and it lies below chance line.
}
}
\description{
\code{calc_curve_shape()} calculates ROC curve shape over a prespecified region.
The prespecified region can be either a range of values in $TPR$ axis, or in
$FPR$ axis.
}
