% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_points.R
\name{calc_partial_roc_points}
\alias{calc_partial_roc_points}
\title{Calculate ROC curve partial points}
\usage{
calc_partial_roc_points(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  lower_threshold,
  upper_threshold,
  ratio
)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation or
\emph{Gold Standard}. For more info on how to select class of interest see
\emph{Methods} below.}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}

\item{lower_threshold, upper_threshold}{Two numbers between 0 and 1, which
indicate lower and upper limits for the region in which to calculate ROC
partial points.}

\item{ratio}{If \code{"tpr"} only points with a TPR in delimited region will be
calculated, if \code{"fpr"} only points with a FPR in the region will do.}
}
\value{
A tibble with two columns: "partial_fpr" and "partial_tpr", which contain
FPR and TPR values for each point in ROC in specified region.
}
\description{
Calculates a series of pairs of (FPR, TPR) points which correspond to ROC
curve points in the specified region. x axis will correspond to
"false positive ratio", while y axis the "true positive ratio".
}
\section{Methods}{

When performing calculations response will be transformed internally to a
factor with levels (0,1), which represent absence or presence of the class we
want to predict. In order to select which class is the one to predict among
availables in response, function follows some criteria according to var type:
\itemize{
\item integer. When working with an integer vector, function will consider the
smallest as the class to predict.
\item character. When working with a character vector, function will consider the
first value after using \code{\link[=sort]{sort()}} over all posible options.
\item factor. When working with a factor var, function will select first class in
\code{\link[=levels]{levels()}}.
}

Al other classes which are not identified as the class to predict will be
collapsed into 0. For this reason, try to provide convenient names when
working or alternatively provide a factor where levels are equal to 0 and 1,
being 1 cases with the condition of interest.
}

\section{Data masking variables}{

Both response and predictor support \code{data-masking} variables. For more
information please check \emph{General usage} in
\link[rlang:args_data_masking]{this article}.

Nevertheless, function may use environment variables as well. In order to use
this type of variables simply set \code{data = NULL} and provide response and
predictor as normal, whithout data-masking.
}

