% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{add_fpauc_partially_proper_lower_bound}
\alias{add_fpauc_partially_proper_lower_bound}
\alias{add_fpauc_concave_lower_bound}
\alias{fpauc_lower_bounds}
\alias{add_fpauc_lower_bound}
\title{Add fpauc lower bound to a ROC plot}
\usage{
add_fpauc_partially_proper_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)

add_fpauc_concave_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)

add_fpauc_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation or
\emph{Gold Standard}. For more info on how to select class of interest see
\emph{Methods} below.}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}

\item{threshold}{A number between 0 and 1, representing lower TPR for the
region of interest.}
}
\description{
Calculate and plot lower bound defined by FpAUC sensitivity index.
}
\details{
\code{add_fpauc_lower_bound()} provides a upper level function which automatically
calculates curve shape and plots a lower bound that ajust to it.

To enforce specific bounds lower level functions
\code{add_fpauc_partially_proper_lower_bound} and \code{add_fpauc_concave_lower_bound}
are provided, which plot lower bound when curve shape partially proper (
presents some kind of hook) or is concave in region of interest.
}
