% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{hide_legend}
\alias{hide_legend}
\title{Hide legend in a ROC plot}
\usage{
hide_legend()
}
\value{
A ggplot theme object.
}
\description{
Hide legend showing name of ploted classifiers and bounds in a
ROC curve plot.
}
