% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{npauc_lower_bounds}
\alias{npauc_lower_bounds}
\alias{add_npauc_lower_bound}
\alias{add_npauc_normalized_lower_bound}
\title{Add NpAUC lower bound to a ROC plot}
\usage{
add_npauc_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)

add_npauc_normalized_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  threshold
)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{fpr, tpr}{Numeric vectors representing FPR and TPR for the classifier.
These arguments will be used when no data, response and predictor are
supplied.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation or
\emph{Gold Standard}. For more info on how to select class of interest see
\emph{Methods} below.}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}

\item{threshold}{A number between 0 and 1 representing lower TPR for the
region of interest.}
}
\description{
Calculate and plot lower bound defined by NpAUC specificity index.

\code{add_npauc_normalized_lower_bound()} allows to plot normalized
lower bound, which is used to formally calculate NpAUC.

\code{add_npauc_lower_bound()} provides a way to plot lower bound previous to
normalization and consecuently corresponds to a lower level function.
}
\examples{
plot_roc_curve(iris, response = Species, predictor = Sepal.Width) +
  add_npauc_lower_bound(
    iris,
    response = Species,
    predictor = Sepal.Width,
    threshold = 0.9
  )
}
