% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc_points}
\alias{plot_roc_points}
\title{Plot classifier points of a ROC curve}
\usage{
plot_roc_points(
  data,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  .condition = NULL
)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{fpr, tpr}{Numeric vectors representing FPR and TPR for the classifier.

These arguments may be used when neither of \code{data}, \code{response} and
\code{predictor} are supplied.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the prediction outcome on each observation
(\emph{Gold Standard}).

If the variable presents more than two possible outcomes, classes or
categories:
\itemize{
\item The outcome of interest (the one to be predicted) will remain distinct.
\item All other categories will be combined into a single category.
}

New combined category represents the "absence" of the condition to predict.
See \code{.condition} for more information.}

\item{predictor}{A data variable which must be numeric, representing values
of a classifier or predictor for each observation.}

\item{.condition}{A value from response that represents class, category or
condition of interest which wants to be predicted.

If \code{NULL}, condition of interest will be selected automatically depending on
\code{response} type.

Once the class of interest is selected, rest of them will be collapsed in a
common category, representing the "absence" of the condition to be predicted.

See \code{vignette("selecting-condition")} for further information on how
automatic selection is performed and details on selecting the condition of
interest.}
}
\description{
Create an scatter plot using ROC curve points.
}
\examples{
plot_roc_points(iris, response = Species, predictor = Sepal.Width)
}
