% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prost.R
\docType{data}
\name{prost}
\alias{prost}
\title{Prostate cancer gene expression data}
\format{
A tibble with 554 observations and 2654 variables:
\describe{
\item{ENSG...}{Gene expression levels. Column names correspond to the
measured gene identifier.}
\item{gleason_score}{Score derived from tissue observation, which indicates
disease severity and progression.}
\item{disease}{Categorical variable indicating whether the sample comes
from diseased ("1") or healthy ("0") tissue.}
\item{prognostic}{Categorical variable indicating whether a poor ("1") or a
good ("0") prognosis is expected for the patient. Diseased cases in the
dataset are assumed to have a poor prognosis when their Gleason score
is equal to or above 8. Non-diseased cases are labelled as "Normal".}
}
}
\source{
Raw data obtained from The Cancer Genome Atlas (TCGA) through the
Genomic Data Commons Data Portal (\url{https://portal.gdc.cancer.gov/}).
Data have been preprocessed and curated by the authors (e.g., gene selection
and variable creation, etc.) to create the final dataset.

The data shown here are in whole or part based upon data
generated by the TCGA Research Network: \url{https://www.cancer.gov/tcga}.
}
\usage{
prost
}
\description{
This dataset contains gene expression levels obtained from healthy and
diseased tissue samples from patients with prostate cancer. The data
includes the expression values for each selected gene, as well as clinical
variables derived from the direct observation of the tissue samples.
}
\details{
Gene identifiers used in columns (e.g. \code{ENSG00000113924},
\code{ENSG00000109182}, ...) correspond to Ensembl gene identifiers.

Gene expression values are reported as \emph{Transcript Per Million} (TPM).
}
\keyword{datasets}
