% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_auc.R, R/np_auc.R
\name{sensitivity_indexes}
\alias{sensitivity_indexes}
\alias{fp_auc}
\alias{np_auc}
\title{Sensitivity indexes}
\usage{
fp_auc(data = NULL, response, predictor, lower_tpr)

np_auc(data = NULL, response, predictor, lower_tpr)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation or
\emph{Gold Standard}. For more info on how to select class of interest see
\emph{Methods} below.}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}

\item{lower_tpr}{Lower value of TPR for the region in which to calculate the
partial area under curve.}
}
\value{
A numeric indicating the partial area under curve in the selected region.
}
\description{
Sensitivity indexes provide different ways of calculating partial area under
ROC curve in a specific TPR range. Two different approaches to calculate this
area are available:
\itemize{
\item \code{fp_aup()} applies \emph{fitted partial area under curve} index (FpAUC). This
one calculates area under curve adjusting to points defined by curve in
selected region.
\item \code{np_auc()} applies \emph{normalized partial area under curve} index (NpAUC),
which calculates area under curve over the whole specified region.
}
}
\section{Methods}{

When performing calculations response will be transformed internally to a
factor with levels (0,1), which represent absence or presence of the class we
want to predict. In order to select which class is the one to predict among
availables in response, function follows some criteria according to var type:
\itemize{
\item integer. When working with an integer vector, function will consider the
smallest as the class to predict.
\item character. When working with a character vector, function will consider the
first value after using \code{\link[=sort]{sort()}} over all posible options.
\item factor. When working with a factor var, function will select first class in
\code{\link[=levels]{levels()}}.
}

Al other classes which are not identified as the class to predict will be
collapsed into 0. For this reason, try to provide convenient names when
working or alternatively provide a factor where levels are equal to 0 and 1,
being 1 cases with the condition of interest.
}

\section{Data masking variables}{

Both response and predictor support \code{data-masking} variables. For more
information please check \emph{General usage} in
\link[rlang:args_data_masking]{this article}.

Nevertheless, function may use environment variables as well. In order to use
this type of variables simply set \code{data = NULL} and provide response and
predictor as normal, whithout data-masking.
}

\examples{
NULL
}
\references{
Franco M. y Vivo J.-M. Evaluating the Performances of Biomarkers over a
Restricted Domain of High Sensitivity. \emph{Mathematics} 9, 2826 (2021).

Jiang Y., Metz C. E. y Nishikawa R. M. A receiver operating characteristic
partial area index for highly sensitive diagnostic tests. \emph{Radiology}
201, 745-750 (1996).
}
