% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_auc.R, R/tp_auc.R
\name{sp_auc}
\alias{sp_auc}
\alias{specificity_indexes}
\alias{tp_auc}
\title{Specificity indexes}
\usage{
sp_auc(data = NULL, response, predictor, lower_fpr, upper_fpr)

tp_auc(data = NULL, response, predictor, lower_fpr, upper_fpr)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation or
\emph{Gold Standard}. For more info on how to select class of interest see
\emph{Methods} below.}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}

\item{lower_fpr, upper_fpr}{Lower and upper values of FPR for the region
in which to calculate the partial area under curve.}
}
\value{
A numeric indicating the partial area under curve in the selected region.
}
\description{
Specifity indexes provide different ways of calculating partial area under
ROC curve in a specific FPR range. Two different approaches to calculate this
area are available:
\itemize{
\item \code{sp_aup()} applies \emph{standardized partial area under curve} index (SpAUC).
This one calculates area under curve adjusting to points defined by curve in
selected region.
\item \code{tp_auc()} applies \emph{tighter partial area under curve} index (TpAUC),
which calculates area under curve over the whole specified region.
}
}
\section{Methods}{

When performing calculations response will be transformed internally to a
factor with levels (0,1), which represent absence or presence of the class we
want to predict. In order to select which class is the one to predict among
availables in response, function follows some criteria according to var type:
\itemize{
\item integer. When working with an integer vector, function will consider the
smallest as the class to predict.
\item character. When working with a character vector, function will consider the
first value after using \code{\link[=sort]{sort()}} over all posible options.
\item factor. When working with a factor var, function will select first class in
\code{\link[=levels]{levels()}}.
}

Al other classes which are not identified as the class to predict will be
collapsed into 0. For this reason, try to provide convenient names when
working or alternatively provide a factor where levels are equal to 0 and 1,
being 1 cases with the condition of interest.
}

\section{Data masking variables}{

Both response and predictor support \code{data-masking} variables. For more
information please check \emph{General usage} in
\link[rlang:args_data_masking]{this article}.

Nevertheless, function may use environment variables as well. In order to use
this type of variables simply set \code{data = NULL} and provide response and
predictor as normal, whithout data-masking.
}

\examples{
NULL
}
\references{
McClish D. K. Analyzing a Portion of the ROC Curve. \emph{Medical Decision Making}
9, 190-195 (1989).

Vivo J.-M., Franco M. y Vicari D. Rethinking an ROC partial area index for
evaluating the classification performance at a high specificity range.
\emph{Advances in Data Analysis and Classification} 12, 683-704 (2018).
}
