% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sumexp_to_df}
\alias{sumexp_to_df}
\title{Transform data in a SummarizedExperiment to a data.frame}
\usage{
sumexp_to_df(se, .n = NULL)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{.n}{An integer or string, representing the index or name of the assay
to use. Same as \code{i} in \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::assay()}} function.

By default, function combines every assay in \code{se} argument.}
}
\value{
A data.frame created from combining assays and colData in a
SummarizedExperiment.
}
\description{
Transforms a SummarizedExperiment into a data.frame which can be used as
input for other functions.
}
