% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{calc_curve_shape}
\alias{calc_curve_shape}
\title{Calculate curve shape over an specific region}
\usage{
calc_curve_shape(
  data = NULL,
  response = NULL,
  predictor = NULL,
  lower_threshold,
  upper_threshold,
  ratio
)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation or
\emph{Gold Standard}. For more info on how to select class of interest see
\emph{Methods} below.}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}

\item{lower_threshold, upper_threshold}{Ranges in which to calculate curve
shape, they can take values from 0 to 1.}

\item{ratio}{If \code{"tpr"}, shape will calculated over TPR ratio. Otherwise, if
\code{"fpr"} it will be calculated over FPR ratio.}
}
\value{
A character indicating ROC curve shape in the specified region. Result
can take any of the following values:
\itemize{
\item \code{Concave}. ROC curve is concave over the entire region.
\item \verb{Partially proper}. ROC curve loses concavity at some point of the
region.
\item \verb{Hook under chance}. ROC curve loses concavity at some point of the
region and it lies below chance line.
}
}
\description{
\code{calc_curve_shape()} calculates ROC curve shape over a prespecified region.
The prespecified region can be either a range of values in $TPR$ axis, or in
$FPR$ axis.
}
\section{Data masking variables}{

Both response and predictor support \code{data-masking} variables. For more
information please check \emph{General usage} in
\link[rlang:args_data_masking]{this article}.

Nevertheless, function may use environment variables as well. In order to use
this type of variables simply set \code{data = NULL} and provide response and
predictor as normal, whithout data-masking.
}

\examples{
NULL
}
