% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{add_fpr_threshold_line}
\alias{add_fpr_threshold_line}
\alias{add_tpr_threshold_line}
\alias{plot_thresholds}
\alias{add_threshold_line}
\title{Add a threshold line to a ROC plot}
\usage{
add_fpr_threshold_line(threshold)

add_tpr_threshold_line(threshold)

add_threshold_line(threshold, ratio = NULL)
}
\arguments{
\item{threshold}{Axis position where to display threshold. Must be a numeric
between 0 and 1.}

\item{ratio}{Ratio in which to display the threshold. If \code{"tpr"} threshold
will be displayed in TPR axis, if \code{"fpr"} it will be displayed in FPR axis.}
}
\description{
Include a threshold line in an specified axis.
}
