% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_points.R
\name{roc_points}
\alias{roc_points}
\title{Calculate curve ROC points}
\usage{
roc_points(data = NULL, response, predictor)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation (Gold Standard),
(see details).}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}
}
\value{
A tibble, with two columns: "tpr", which contains values for "true positive
ratio" or y axis, and "fpr", which contains values for "false positive ratio"
or x axis.
}
\description{
Calculates a series of pairs of (FPR, TPR) points which correspond to ROC
curve displayed points. X axis will represent "false positive ratio", while Y
axis the "true positive ratio".
}
\details{
When performing calculations response will be transformed internally to a
factor with levels (0,1), which represent absence or presence of the class we
want to predict. In order to select which class is the one to predict among
availables in response, function follows some criteria according to var type:
\itemize{
\item integer. When working with an integer vector, function will consider the
smallest as the class to predict.
\item character. When working with a character vector, function will consider the
first value after using \code{\link[=sort]{sort()}} over all posible options.
\item factor. When working with a factor var, function will select first class in
\code{\link[=levels]{levels()}}.
}

Al other classes which are not identified as the class to predict will be
collapsed into 0. For this reason, try to provide convenient names when
working or alternatively provide a factor where levels are equal to 0 and 1,
being 1 cases with the condition of interest.
}
\section{Data Masking variables}{

Both response and predictor support \code{data-masking} variables. For more
information please check \link[rlang:args_data_masking]{rlang::args_data_masking}.

Nevertheless, function may use environment variables as well. In order to use
this type of variables simply set \code{data = NULL} and provide response and
predictor as normal.
}

\examples{
NULL
}
