% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{spauc_lower_bounds}
\alias{spauc_lower_bounds}
\alias{add_spauc_lower_bound}
\title{Add SpAUC lower bound to a ROC plot}
\usage{
add_spauc_lower_bound(
  data = NULL,
  fpr = NULL,
  tpr = NULL,
  response = NULL,
  predictor = NULL,
  lower_threshold,
  upper_threshold
)
}
\arguments{
\item{data}{A data.frame or extension (e.g. a tibble) containing values for
predictors and response variables.}

\item{response}{A data variable which must be a factor, integer or character
vector representing the class to predict on each observation or
\emph{Gold Standard}. For more info on how to select class of interest see
\emph{Methods} below.}

\item{predictor}{A data variable wich must be numeric, representing values of
a classifier or predictor for each observation.}

\item{lower_threshold, upper_threshold}{Two numbers between 0 and 1,
representing lower and upper FPR for a region of interest.}
}
\description{
Calculate and plot lower bound defined by SpAUC specificity index.
}
\details{
SpAUC presents some limitations regarding its lower bound. The bound defined
by this index cannot be applied to sections where ROC curve is defined under
chance line.

\code{add_spauc_lower_bound()} doesn't make any check to ensure the index can be
safely applied, consequently it allows to enforce representation even though
SpAUC cound't be calculated in the region.
}
